more cleanup and some mypy fixes
