+++ mypy-baseline.txt
@@ -676,14 +676,8 @@ posthog/tasks/warehouse.py:0: error: Item "None" of "Any | None" has no attribut
 posthog/tasks/warehouse.py:0: error: Value of type "Any | None" is not indexable  [index]
 posthog/temporal/data_imports/pipelines/pipeline/hogql_schema.py:0: error: "Array[Any]" has no attribute "as_py"  [attr-defined]
 posthog/temporal/data_imports/pipelines/pipeline/test/test_pipeline_utils.py:0: error: Argument 1 to "schema" has incompatible type "list[object]"; expected "Iterable[Field[Any]] | Iterable[tuple[str, DataType]] | Mapping[str, DataType]"  [arg-type]
-posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: "Array[Any]" has no attribute "as_py"  [attr-defined]
-posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: "Array[Any]" has no attribute "as_py"  [attr-defined]
 posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: Argument 1 to "Decimal" has incompatible type "Any | None"; expected "Decimal | float | str | tuple[int, Sequence[int], int]"  [arg-type]
 posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: Argument 1 to "__call__" of "_UFunc_Nin1_Nout1" has incompatible type "Any | None"; expected "int | float | complex | str | bytes | generic"  [arg-type]
-posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: Argument 3 to "set_column" of "Table" has incompatible type "ChunkedArray[Any]"; expected "Array[Any] | list[Any]"  [arg-type]
-posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: Argument 3 to "set_column" of "Table" has incompatible type "ChunkedArray[Scalar[TimestampType[Literal['us'], None]]]"; expected "Array[Any] | list[Any]"  [arg-type]
-posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: List comprehension has incompatible type List[str | None]; expected List[bool | None]  [misc]
-posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: No overload variant of "array" matches argument types "list[None]", "object"  [call-overload]
 posthog/temporal/data_imports/pipelines/pipeline/utils.py:0: error: Statement is unreachable  [unreachable]
 posthog/temporal/data_imports/pipelines/pipeline_sync.py:0: error: Incompatible types in assignment (expression has type "dict[str, dict[str, str | bool]] | dict[str, str]", variable has type "dict[str, dict[str, str]]")  [assignment]
 posthog/temporal/data_imports/pipelines/pipeline_sync.py:0: error: Incompatible types in assignment (expression has type "object", variable has type "DataWarehouseCredential | Combinable | None")  [assignment]
