+++ posthog/temporal/data_imports/pipelines/pipeline/utils.py
@@ -4,7 +4,7 @@ import uuid
 import decimal
 import hashlib
 import datetime
-from collections.abc import Iterator, Sequence
+from collections.abc import Iterator
 from ipaddress import IPv4Address, IPv6Address
 from typing import TYPE_CHECKING, Any, cast
 
@@ -14,10 +14,7 @@ import pyarrow as pa
 import deltalake as deltalake
 import pyarrow.compute as pc
 from dateutil import parser
-from dlt.common.data_types.typing import TDataType
-from dlt.common.libs.deltalake import ensure_delta_compatible_arrow_schema
 from dlt.common.normalizers.naming.snake_case import NamingConvention
-from dlt.sources import DltResource
 from structlog.types import FilteringBoundLogger
 
 from posthog.temporal.data_imports.pipelines.pipeline.consts import PARTITION_KEY
@@ -29,17 +26,6 @@ if TYPE_CHECKING:
 # deltalake maximum precision and scale
 DEFAULT_PYARROW_DECIMAL_TYPE = pa.decimal128(38, 32)
 DEFAULT_PARTITION_TARGET_SIZE_IN_BYTES = 200 * 1024 * 1024  # 200 MB
-DLT_TO_PA_TYPE_MAP = {
-    "text": pa.string(),
-    "bigint": pa.int64(),
-    "bool": pa.bool_(),
-    "timestamp": pa.timestamp("us"),
-    "json": pa.string(),
-    "double": pa.float64(),
-    "date": pa.date64(),
-    "time": pa.timestamp("us"),
-    "decimal": pa.float64(),
-}
 DATE_FORMAT_BY_PARTITION_FORMAT = {
     "hour": "%Y-%m-%dT%H",
     "day": "%Y-%m-%d",
@@ -88,39 +74,6 @@ def safe_parse_datetime(date_str) -> None | pa.TimestampScalar | datetime.dateti
         return None
 
 
-def _get_primary_keys(resource: DltResource) -> list[str] | None:
-    primary_keys = resource._hints.get("primary_key")
-    if primary_keys is None:
-        return None
-    if isinstance(primary_keys, str):
-        return [normalize_column_name(primary_keys)]
-    if isinstance(primary_keys, list | Sequence):
-        return [normalize_column_name(pk) for pk in primary_keys]
-    raise Exception(f"primary_keys of type {primary_keys.__class__.__name__} are not supported")
-
-
-def _get_column_hints(resource: DltResource) -> dict[str, TDataType | None] | None:
-    columns = resource._hints.get("columns")
-    if columns is None:
-        return None
-    return {key: value.get("data_type") for key, value in columns.items()}  # type: ignore
-
-
-def _handle_null_columns_with_definitions(table: pa.Table, source: SourceResponse) -> pa.Table:
-    column_hints = source.column_hints
-    if column_hints is None:
-        return table
-    for field_name, data_type in column_hints.items():
-        if data_type is None:
-            continue
-        normalized_field_name = normalize_column_name(field_name)
-        # If the table doesn't have all fields, then add a field with all Nulls and the correct field type
-        if normalized_field_name not in table.schema.names:
-            new_column = pa.array([None] * table.num_rows, type=DLT_TO_PA_TYPE_MAP[data_type])
-            table = table.append_column(normalized_field_name, new_column)
-    return table
-
-
 def get_default_value_for_pyarrow_type(arrow_type: pa.DataType) -> Any:
     if pa.types.is_integer(arrow_type):
         return 0
@@ -152,113 +105,6 @@ def get_default_value_for_pyarrow_type(arrow_type: pa.DataType) -> Any:
         return None
     else:
         raise ValueError(f"Unsupported PyArrow type: {arrow_type}")
-
-
-def _evolve_pyarrow_schema(table: pa.Table, delta_schema: deltalake.Schema | None) -> pa.Table:
-    py_table_field_names = table.schema.names
-
-    for column_name in table.column_names:
-        column = table.column(column_name)
-        field = table.field(column_name)
-
-        # Change pa.structs to JSON string
-        if pa.types.is_struct(column.type) or pa.types.is_list(column.type):
-            json_column = pa.array([_json_dumps(row.as_py()) if row.as_py() is not None else None for row in column])
-            table = table.set_column(table.schema.get_field_index(column_name), column_name, json_column)
-            column = table.column(column_name)
-        # Change pa.duration to int with total seconds
-        elif pa.types.is_duration(column.type):
-            seconds_column = pa.array(
-                [row.as_py().total_seconds() if row.as_py() is not None else None for row in column]
-            )
-            table = table.set_column(table.schema.get_field_index(column_name), column_name, seconds_column)
-            column = table.column(column_name)
-
-        # Convert nanosecond timestamps to microseconds and convert to UTC
-        if pa.types.is_timestamp(field.type) and (field.type.unit == "ns" or field.type.tz is not None):
-            microsecond_timestamps = pc.cast(column, pa.timestamp("us"), safe=False)
-            table = table.set_column(table.schema.get_field_index(column_name), column_name, microsecond_timestamps)
-
-    if delta_schema:
-        for field in delta_schema.to_pyarrow():
-            if field.name not in py_table_field_names:
-                if field.nullable:
-                    new_column_data = pa.array([None] * table.num_rows, type=field.type)
-                else:
-                    new_column_data = pa.array(
-                        [get_default_value_for_pyarrow_type(field.type)] * table.num_rows, type=field.type
-                    )
-                table = table.append_column(field, new_column_data)
-
-            # If the delta table schema has a larger scale/precision, then update the
-            # pyarrow schema to use the larger values so that we're not trying to downscale
-            if isinstance(field.type, pa.Decimal128Type):
-                py_arrow_table_column = table.column(field.name)
-
-                if isinstance(py_arrow_table_column.type, pa.Decimal128Type) and (
-                    field.type.precision > py_arrow_table_column.type.precision
-                    or field.type.scale > py_arrow_table_column.type.scale
-                ):
-                    field_index = table.schema.get_field_index(field.name)
-                    new_decimal_type = pa.decimal128(field.type.precision, field.type.scale)
-
-                    new_schema = table.schema.set(
-                        field_index,
-                        table.schema.field(field_index).with_type(new_decimal_type),
-                    )
-                    table = table.cast(new_schema)
-
-            # If the deltalake schema has a different type to the pyarrows table, then cast to the deltalake field type
-            py_arrow_table_column = table.column(field.name)
-            if field.type != py_arrow_table_column.type:
-                if isinstance(field.type, pa.TimestampType):
-                    # If different timezones, cast to the correct tz
-                    if (
-                        isinstance(py_arrow_table_column.type, pa.TimestampType)
-                        and field.type.tz != py_arrow_table_column.type.tz
-                    ):
-                        casted_column = table.column(field.name).cast(field.type)
-                        table = table.set_column(
-                            table.schema.get_field_index(field.name),
-                            field.name,
-                            casted_column.combine_chunks(),
-                        )
-                    else:
-                        timestamp_array = pa.array(
-                            [safe_parse_datetime(s) for s in table.column(field.name)], type=field.type
-                        )
-                        table = table.set_column(
-                            table.schema.get_field_index(field.name),
-                            field.name,
-                            timestamp_array,
-                        )
-                else:
-                    table = table.set_column(
-                        table.schema.get_field_index(field.name),
-                        field.name,
-                        table.column(field.name).cast(field.type),
-                    )
-
-                py_arrow_table_column = table.column(field.name)
-
-            py_arrow_table_field = table.field(field.name)
-            # If the deltalake schema expects no nulls, but the pyarrow schema is nullable, then fill the nulls
-            if not field.nullable and py_arrow_table_field.nullable:
-                filled_nulls_arr = py_arrow_table_column.fill_null(
-                    fill_value=get_default_value_for_pyarrow_type(py_arrow_table_field.type)
-                )
-                table = table.set_column(
-                    table.schema.get_field_index(field.name), field, filled_nulls_arr.combine_chunks()
-                )
-
-    # Change types based on what deltalake tables support
-    return table.cast(ensure_delta_compatible_arrow_schema(table.schema))
-
-
-def _append_debug_column_to_pyarrows_table(table: pa.Table, load_id: int) -> pa.Table:
-    debug_info = f'{{"load_id": {load_id}}}'
-    column = pa.array([debug_info] * table.num_rows, type=pa.string())
-    return table.append_column("_ph_debug", column)
 
 
 def normalize_table_column_names(table: pa.Table) -> pa.Table:
@@ -426,7 +272,7 @@ def _convert_uuid_to_string(row: dict) -> dict:
     return {key: str(value) if isinstance(value, uuid.UUID) else value for key, value in row.items()}
 
 
-def _json_dumps(obj: Any) -> str:
+def json_dumps_3000(obj: Any) -> str:
     try:
         return orjson.dumps(obj).decode()
     except TypeError:
@@ -731,7 +577,7 @@ def _process_batch(table_data: list[dict], schema: pa.Schema | None = None) -> p
         if len(unique_types_in_column) > 1 and str in unique_types_in_column and dict in unique_types_in_column:
             json_array = pa.array(
                 [
-                    None if s is None else _json_dumps(s) if isinstance(s, dict | list) else s
+                    None if s is None else json_dumps_3000(s) if isinstance(s, dict | list) else s
                     for s in _to_list_array(columnar_table_data[field_name])
                 ]
             )
@@ -744,7 +590,7 @@ def _process_batch(table_data: list[dict], schema: pa.Schema | None = None) -> p
         # If there are multiple types that aren't a list, then JSON stringify everything
         if len(unique_types_in_column) > 1:
             json_array = pa.array(
-                [None if s is None else _json_dumps(s) for s in _to_list_array(columnar_table_data[field_name])]
+                [None if s is None else json_dumps_3000(s) for s in _to_list_array(columnar_table_data[field_name])]
             )
             columnar_table_data[field_name] = json_array
             py_type = str
@@ -755,7 +601,7 @@ def _process_batch(table_data: list[dict], schema: pa.Schema | None = None) -> p
         # Convert any dict/lists to json strings to avoid schema mismatches in nested objects
         if issubclass(py_type, dict | list):
             json_str_array = pa.array(
-                [None if s is None else _json_dumps(s) for s in _to_list_array(columnar_table_data[field_name])]
+                [None if s is None else json_dumps_3000(s) for s in _to_list_array(columnar_table_data[field_name])]
             )
             columnar_table_data[field_name] = json_str_array
             py_type = str
